+++ ChangeLog.md
@@ -126,6 +126,7 @@ These artifacts include extensions for the types available in the latter JDKs, s
 - [`KT-8867`](https://youtrack.jetbrains.com/issue/KT-8867) Rename: Rename all overridden members if user chooses to refactor base declaration(s)
 - Pull Up: Drop 'override' modifier if moved member doesn't override anything
 - [`KT-13660`](https://youtrack.jetbrains.com/issue/KT-13660) Move: Do not drop object receivers when calling variable of extension functional type
+- [`KT-13759`](https://youtrack.jetbrains.com/issue/KT-13759) Rename: Process object-wrapping alias references
 
 ##### New features